+++ Assets/Scenes/TutorialScene/LightingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e492e0e7a9754414c9aaee05f6984a01
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 112000000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
