+++ Assets/Scripts/Train/StartTrain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4cbb7d39659830941843177fa68c1620
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
