+++ Assets/Scripts/Train/TutorialTrain.cs
@@ -5,15 +5,13 @@ using UnityEngine;
 public class TutorialTrain : Train
 {
     [SerializeField] private GameObject _tutorialObjectInTrainPrefab; // 기차 내 오브젝트 프리팹이 모두 있는 배열
-    [SerializeField] private GameObject _treasureBox;
-    [SerializeField] private Transform _treasureBoxSpawnPoint;
 
     private int _enemyCount;
-    private bool _isClear;
 
     public override void Start()
     {
         base.Start();
+        Init();
     }
 
     /// <summary>
@@ -30,16 +28,6 @@ public class TutorialTrain : Train
         system.transform.parent = transform;
         _enemyCount = system.EnemyCount;
     }
-    private void ClearStage()
-    {
-        if (_isClear == true)
-            return;
-        _isClear = true;
-
-        OpenDoor();
-
-        Instantiate(_treasureBox, _treasureBoxSpawnPoint.position, Quaternion.identity); // 상자 생성 
-    }
 
     /// <summary>
     /// 적이 죽었을 때 실행되는 함수
@@ -51,15 +39,7 @@ public class TutorialTrain : Train
         // 적 카운트가 0이 되면 게임 클리어
         if (_enemyCount <= 0)
         {
-            ClearStage();
-        }
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            InGameManager.Instance.NextStage();
+            Init();
         }
     }
 }
